+++ webui/react/src/hooks/useModal/Checkpoint/useModalCheckpoint.tsx
@@ -5,10 +5,10 @@ import React, { useCallback, useEffect, useMemo } from 'react';
 import Badge, { BadgeType } from 'components/Badge';
 import HumanReadableNumber from 'components/HumanReadableNumber';
 import Link from 'components/Link';
-import useModal, { ModalCloseReason, ModalHooks } from 'hooks/useModal/useModal';
 import { paths } from 'routes/utils';
 import { detApi } from 'services/apiConfig';
 import { readStream } from 'services/utils';
+import useModal, { ModalCloseReason, ModalHooks } from 'shared/hooks/useModal/useModal';
 import { formatDatetime } from 'shared/utils/datetime';
 import { humanReadableBytes } from 'shared/utils/string';
 import {
