+++ webui/react/src/hooks/useModal/Model/useModalModelCreate.tsx
@@ -6,10 +6,10 @@ import Link from 'components/Link';
 import EditableMetadata from 'components/Metadata/EditableMetadata';
 import EditableTagList from 'components/TagList';
 import useModalCheckpointRegister from 'hooks/useModal/Checkpoint/useModalCheckpointRegister';
-import useModal, { ModalHooks as Hooks } from 'hooks/useModal/useModal';
 import usePrevious from 'hooks/usePrevious';
 import { paths } from 'routes/utils';
 import { getModels, postModel } from 'services/api';
+import useModal, { ModalHooks as Hooks } from 'shared/hooks/useModal/useModal';
 import { clone, isEqual } from 'shared/utils/data';
 import { ErrorType } from 'shared/utils/error';
 import { Metadata } from 'types';
